+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Add legal entry to the dashboard (!1020)
 - Add a new call_in field to the room field of events (!999)
 - Hide dial-in option when no call-in info is sent (!1001)
+- Glitchtip integration (!1016)
 
 ## 1.5.x